requirements.txt
gradio
torch
transformers
sentencepiece
opencv-python-headless
pillow
numpy
accelerate